payment/settlement 금액 타입 BigDecimal 통일 및 ai 문서 정합화

+++ payment/src/main/java/com/example/payment/application/dto/ChargeConfirmResult.java
@@ -1,14 +1,15 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.ChargeStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record ChargeConfirmResult(
         UUID chargeId,
         ChargeStatus chargeStatus,
-        Long approvedAmount,
-        Long walletBalance,
+        BigDecimal approvedAmount,
+        BigDecimal walletBalance,
         LocalDateTime approvedAt
 ) {
 }
