+++ payment/src/main/java/com/example/payment/application/dto/PaymentRefundResult.java
@@ -2,6 +2,7 @@ package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.PaymentRefundStatus;
 import com.example.payment.domain.enumtype.PaymentRefundType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -12,7 +13,7 @@ public record PaymentRefundResult(
         UUID orderCancelRequestId,
         PaymentRefundStatus refundStatus,
         PaymentRefundType refundType,
-        Long totalRefundAmount,
+        BigDecimal totalRefundAmount,
         List<PaymentRefundItemResult> itemResults,
         LocalDateTime processedAt
 ) {
