+++ payment/src/main/java/com/example/payment/application/dto/WalletSummaryResult.java
@@ -1,5 +1,6 @@
 package com.example.payment.application.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ import java.util.UUID;
 public record WalletSummaryResult(
         UUID walletId,
         UUID memberId,
-        Long balance,
+        BigDecimal balance,
         LocalDateTime updatedAt
 ) {
 }
