+++ payment/src/main/java/com/example/payment/domain/service/OrderPaymentValidationGateway.java
@@ -1,8 +1,9 @@
 package com.example.payment.domain.service;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface OrderPaymentValidationGateway {
 
-    OrderPaymentValidationData validate(UUID orderId, UUID buyerId, Long amount);
+    OrderPaymentValidationData validate(UUID orderId, UUID buyerId, BigDecimal amount);
 }
