+++ payment/src/main/java/com/example/payment/presentation/dto/response/EscrowTransactionItemResponse.java
@@ -2,6 +2,7 @@ package com.example.payment.presentation.dto.response;
 
 import com.example.payment.application.dto.EscrowTransactionItemResult;
 import com.example.payment.domain.enumtype.EscrowTransactionType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -13,9 +14,9 @@ public record EscrowTransactionItemResponse(
         UUID sellerMemberId,
         UUID buyerMemberId,
         EscrowTransactionType transactionType,
-        Long amount,
-        Long beforeAmount,
-        Long afterAmount,
+        BigDecimal amount,
+        BigDecimal beforeAmount,
+        BigDecimal afterAmount,
         UUID referenceId,
         String referenceType,
         String description,
@@ -42,4 +43,3 @@ public record EscrowTransactionItemResponse(
         );
     }
 }
-
